Improve AccountManager component
