+++ src/styles/GeneralStyles.css
